chore: 엔티티별 패키지 구조 분리
