+++ src/main/java/com/eodya/api/place/domain/PlaceStatus.java
@@ -1,4 +1,4 @@
-package com.eodya.api.entity;
+package com.eodya.api.place.domain;
 
 public enum PlaceStatus {
     BLOOMING,
